backend/requirements.txt
django~=5.2
djangorestframework~=3.16
django-cors-headers~=4.3
psycopg2-binary~=2.9
gunicorn~=23.0
whitenoise~=6.11
celery[redis]~=5.5
redis~=5.0
pywebpush~=2.0
djangorestframework-simplejwt~=5.5
django-environ~=0.11
